# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gallery API",
        "description": "Internal HTTP API for the self-hosted media gallery — a **hybrid REST-RPC**\ninterface served under the `/api` base path. All responses are JSON.\n\n**Conventions**\n- Reads return `200` with the resource or collection.\n- Creates return `201` with the created resource.\n- Updates (`PUT`/`PATCH`) return `200` with the updated resource.\n- Deletes return `204` with no body.\n- Non-CRUD actions (login, migrate, scan, bulk operations) are `POST`\n  verbs that return `200` with a small result object.\n- Errors return `{ \"error\": \"<MachineCode>\", \"message\": \"<human text>\" }`\n  with a `4xx`/`5xx` status. `error` is a stable PascalCase code safe to\n  branch on; `message` is for display and may change.\n\n**Auth** — a Bearer token (`Authorization: Bearer <token>`) obtained from\n`POST /auth/login` (valid 24h). All `GET`s are public; state-changing\nmethods require a token, except a small public allowlist (media tagging\nand the batched `POST /media/by-ids` read).",
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
...
        "/duplicates/report": {
            "get": {
                "tags": [
                    "Duplicates"
                ],
                "summary": "Latest duplicate report",
                "description": "GET /duplicates/report — The latest duplicate-scan report, with dismissed\npairs filtered out and each pair enriched with media file names/hashes.",
                "operationId": "cf87d7c7b3fda80a991907a425d0fc0f",
                "responses": {
                    "200": {
                        "description": "The report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DuplicateReport"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NoReportsFound",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/duplicates/scan": {
            "post": {
                "tags": [
                    "Duplicates"
                ],
                "summary": "Run a duplicate scan",
                "description": "POST /duplicates/scan — Run the perceptual-hash duplicate scanner and\nreturn summary stats (images compared, candidates, duplicates found).",
                "operationId": "1723375dae6967be7242b04b58a65e47",
                "responses": {
                    "200": {
                        "description": "Scan stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScanResult"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "ScanFailed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/duplicates/dismissals": {
            "post": {
                "tags": [
                    "Duplicates"
                ],
                "summary": "Dismiss a duplicate pair",
                "description": "POST /duplicates/dismissals — Dismiss a pair of media items as not-duplicates.\nBody: { media_id_1, media_id_2 }.",
                "operationId": "70b30de80982dfb5b34e060d84757c9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "media_id_1",
                                    "media_id_2"
                                ],
                                "properties": {
                                    "media_id_1": {
                                        "type": "integer"
                                    },
                                    "media_id_2": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dismissed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DismissResult"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "InvalidInput",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DismissResult": {
                "description": "A dismissed duplicate pair.",
                "properties": {
                    "dismissed": {
                        "type": "boolean"
                    },
                    "media_id_1": {
                        "type": "integer"
                    },
                    "media_id_2": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "DuplicateMatch": {
                "description": "A duplicate media pair with similarity scores.",
                "properties": {
                    "media_1": {
                        "properties": {
                            "media_id": {
                                "type": "integer"
                            },
                            "file_name": {
                                "type": "string"
                            },
                            "hash": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "media_2": {
                        "properties": {
                            "media_id": {
                                "type": "integer"
                            },
                            "file_name": {
                                "type": "string"
                            },
                            "hash": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "distance": {
                        "description": "Hamming distance between fingerprints.",
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "ssim": {
                        "description": "Structural similarity (SSIM) score.",
                        "type": [
                            "number",
                            "null"
                        ],
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "DuplicateReport": {
                "description": "The latest duplicate-detection report.",
                "properties": {
                    "report_file": {
                        "type": "string"
                    },
                    "generated_at": {
                        "description": "ISO timestamp the report was generated.",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "images_compared": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "duplicates_found": {
                        "type": "integer"
                    },
                    "matches": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DuplicateMatch"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Standard error envelope.",
                "properties": {
                    "error": {
                        "description": "Stable PascalCase machine code (e.g. MediaNotFound).",
                        "type": "string"
                    },
                    "message": {
                        "description": "Human-readable message for display.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ScanResult": {
                "description": "Duplicate-scan summary statistics.",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "images_compared": {
                        "type": "integer"
                    },
                    "lsh_candidates": {
                        "type": "integer"
                    },
                    "duplicates_found": {
                        "type": "integer"
                    },
                    "execution_time": {
                        "description": "Execution time in seconds.",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Admin bearer token issued by POST /auth/login (valid 24 hours).",
                "scheme": "bearer"
            }
        }
    },
...
}